+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 FEATURES:
 
+* **Cluster Packages** Client methods for `cluster_packages` API endpoints
 * **Forms** Client methods for `forms` API endpoints
 
 ## 0.3.6 (November 13, 2023)